:recycle: refactor: reducer 리팩터링
