+++ src/containers/QuantityInput/index.jsx
@@ -4,7 +4,7 @@ import {
   requestExistProductAdd,
   requestExistProductDecrease,
   requestUpdateProductQuantity,
-} from 'modules/cart';
+} from 'modules/cart/cart';
 import { useDispatch, useSelector } from 'react-redux';
 import { InputWrapperStyled, InputStyled, ButtonWrapperStyled } from './style';
