+++ src/modules/reducerUtilities.js
@@ -1,0 +1,8 @@
+const updateObject = (state, newValues) => {
+  return {
+    ...state,
+    ...newValues,
+  };
+};
+
+export default updateObject;
